+++ Starter.Repository.Tests/Repositories/CatRepositoryTests.cs
@@ -44,7 +44,7 @@ namespace Starter.Repository.Tests.Repositories
         [Category("Integration")]
         public async Task Create_Cat_Successful()
         {
-            var cat = new Cat { Id = Guid.NewGuid(), Name = Guid.NewGuid().ToString(), AbilityId = (int)Ability.Napping };
+            var cat = new Cat { Id = Guid.NewGuid(), Name = Guid.NewGuid().ToString(), AbilityId = Ability.Napping };
 
             Cats.Add(cat);
             await CatRepository.CreateAsync(cat);
